// src/shared/ui/components/Charts/types.ts
export enum ChartsPeriodSelectValue {
  LAST_30_DAYS = 'last_30_days',
  LAST_365_DAYS = 'last_365_days',
  LAST_7_DAYS = 'last_7_days',
  LAST_90_DAYS = 'last_90_days',
  MONTH_TO_DATE = 'month_to_date',
  QUARTER_TO_DATE = 'quarter_to_date',
  WEEK_TO_DATE = 'week_to_date',
  YEAR_TO_DATE = 'year_to_date',
}